main_RF_KingsData_ThreeClasses.m
In the given eeg data, only 4 episodes of epilepsy data corresponding to the labelling in the slides could be found. The model was trained using three of the episodes and tested using the remaining episodes. And the test results are plotted as images. The horizontal coordinate of the image is the time axis and the vertical coordinate is the number of channels, the legend contains three items, the number of channels predicted to be pre-seizure, the number predicted to be seizure, and the number predicted to be non-seizure